Update FizzBuzz_2.py
This is without elif statements. Regarding your other comment: I know right?! how can you say the answer is supposed to be "FizzBuzz" and NOT give a string concatenation in the answer. I was tempted to make one with it, but then this somehow was the path of least resistance. :)